requirements.txt
beautifulsoup4==4.12.3
Django==4.2.16
django-appconf==1.0.6
django-bootstrap-v5==1.0.11
django-braces==1.16.0
django-imagekit==5.0.0
misaka==2.1.1
pilkit==3.0
pillow==11.0.0
